Create a game class
